Fix signedness warning in gio/gio-tool.c:attribute_flags_to_string()
gio/gio-tool.c: In function ‘attribute_flags_to_string’:
gio/gio-tool.c:171:17: error: comparison of integer expressions of different signedness: ‘int’ and ‘long unsigned int’
  171 |   for (i = 0; i < G_N_ELEMENTS (flag_descr); i++)
      |                 ^

+++ gio/gio-tool.c
@@ -149,7 +149,7 @@ char *
 attribute_flags_to_string (GFileAttributeInfoFlags flags)
 {
   GString *s;
-  int i;
+  gsize i;
   gboolean first;
   struct {
     guint32 mask;
